requirements.txt
python-dotenv
openai
flask
requests
flask-cors
bcut-asr-branch
bilix
youtube-transcript-api